ipdg: refactor such that only one entry point into the ipdg interface is needed
there are now three entry points into the ipdg interface, but only one
is needed to perform full service, i.e. full population of the caller's
struct nmsg_ipdg object. the three functions are needed since the caller
could be operating on raw frames that need de-framing and IP reassembly,
or could be operating on cooked IP datagrams (no de-framing or
reassembly needed) or could be operating on IP datagrams that need
reassembly.