input: nmsg_input_read_null(): fix invalid attempt to free detached payloads
When a payload is detached, the pointer pointing to it in the nmsg
structure is zeroed out. However, the n_payloads count was not updated to
reflect this. As a result, the call to nmsg__nmsg__free_unpacked would attempt
to free the now detached payloads, which would cause a crash due to a NULL
pointer dereference.

Set n_payloads to 0 to reflect the fact that all payloads have been detached,
and also free the payloads array.

[edmonds@fsi.io: adjust commit message.]