build system: explicit checks and summary messages for documentation building tools
add build system checks for the two documentation building tools,
doxygen and docbook2x-man. these programs may not be available in the
build environment, so explicitly search for them and disable the
documentation targets in the makefile if they are not available. this
makes it possible to build the rest of the nmsg distribution on systems
where these tools are not available.

also print summary messages at the end of the configure run indicating
if the documentation will be built or not.