res: Rewrite nmsg_res_lookup() to use a straight-forward switch
The previous implementation of this code tripped a pedantic warning from
clang:

```
nmsg/res.c:43:10: warning: comparison of constant 16 with expression of type 'enum nmsg_res' is always
      true [-Wtautological-constant-out-of-range-compare]
            val <= (sizeof(res_strings) / sizeof(char *)) &&
            ~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```

Rather than casting val, rewrite the function to use a straight-forward
switch block. This also has the advantage that if we add a new res value
but neglect to update this function, it will throw a -Wswitch
diagnostic.

(Fixes #8.)