nmsgtool: Add -F / --filter option
This commit adds a -F / --filter command-line option that can be
specified multiple times.

The parameter to --filter will be split on the first "," character, if
any. The first token will be treated as the 'name' parameter to
nmsg_io_add_filter_module(), while the token after the "," will be
passed as a NUL-terminated string via the 'param' and 'len_param'
parameters.