Allow message filter functions to replace the input message
Change the 'msg' parameter to functions of type nmsg_filter_message_fp
and type nmsg_fltmod_filter_message_fp to be type pointer to
nmsg_message_t.

This allows filter functions to completely replace the input message
with a newly constructed message. If this is done, however, the filter
function is itself responsible for disposing the old message.