tests/test-layout-fltmod_plugin: Don't break if a reserved field is replaced
Per @hstern's feedback, the previous test_sizeof() would break if we
ever replaced a reserved field with a new field, without updating the
test.

This commit removes the explicit sizeof() checks for every field in the
struct in the "test-sizeof" test. Instead, I added a new "test-offsetof"
test which explicitly references only the non-reserved fields in the
structure, and also, by using the offsetof() macro, verifies the field
ordering as well. (The original test didn't do this.)