nmsg_message_get_num_field_values_by_idx(): Don't crash on virtual fields
Virtual fields don't have a field->descr (since they are not backed by
protobuf fields), so we need to special case the handling, otherwise
we'll crash on a NULL pointer dereference.

Virtual fields are completely synthetic and don't construct their field
content until the field->get function is called, so we need to actually
call the field getter function and retrieve each field value until
failure in order to count the number of field values in a virtual field.