Have ./configure generate the test scripts with correct paths
This is for Issue #15.

Move the three tests.sh to tests.sh.in and add to AC_CONFIG_FILES
and make executable in configure.ac.

Use @abs_top_srcdir@ to refer to correct payload files.
Use @abs_top_builddir@ to use nmsgtool and crc32c_test.

Run tests in the builddir, not srcdir.

Code cleanup in payload-crc32c-tests to actually use
some unused variables (x) that were defined.

Fix second issue in udp-checksum-tests test.sh which
was using the installed, possibly old, version of nmsgtool
instead of the newly-built nmsgtool.