nmsgtool: set only one of stats_output and stats_user
Previously, if nmsgtool were run with a kicker-managed file output and a
socket output, we may have had stats_output set to the socket output and
stats_user set to the kickfile structure of the file output. This would
cause us to print stats for both the file output (due to the stats_user
match) and the socket output (due to the stats_output match).